Avoid type casting warning.

+++ libtiff/tif_compress.c
@@ -147,7 +147,7 @@ _TIFFSetDefaultCompressionState(TIFF* tif)
 int
 TIFFSetCompressionScheme(TIFF* tif, int scheme)
 {
-	const TIFFCodec *c = TIFFFindCODEC(scheme);
+	const TIFFCodec *c = TIFFFindCODEC((uint16) scheme);
 
 	_TIFFSetDefaultCompressionState(tif);
 	/*
